Log out and changing duty status works correctly.
