+++ app/src/main/java/org/wildaid/ofish/ui/home/HomeActivity.kt
@@ -32,12 +32,19 @@ class HomeActivity : AppCompatActivity(R.layout.activity_home) {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        activityViewModel.userEventLiveData.observe(this, EventObserver() {
+        activityViewModel.userEventLiveData.observe(this, EventObserver {
             when (it) {
                 HomeActivityViewModel.UserEvent.AskDutyConfirmationEvent -> askToChangeDuty()
+                HomeActivityViewModel.UserEvent.AskUserLogoutEvent -> askToLogout()
+                HomeActivityViewModel.UserEvent.UserLogoutEvent -> onUserLoggedOut()
+                HomeActivityViewModel.UserEvent.BecomeNotAtSea -> navigateToPatrolSummary()
             }
         })
         activityViewModel.timerLiveData.observe(this, EventObserver(::setOrCancelTimer))
+    }
+
+    private fun navigateToPatrolSummary() {
+        navigation.navigate(R.id.action_profileFragment_to_patrolSummaryFragment2)
     }
 
     override fun onNewIntent(intent: Intent?) {
@@ -69,6 +76,23 @@ class HomeActivity : AppCompatActivity(R.layout.activity_home) {
         ).bundle()
 
         navigation.navigate(R.id.ask_change_duty_dialog, dialogBundle)
+    }
+
+    private fun onUserLoggedOut() {
+        navigation.navigate(R.id.action_profileFragment_to_login_activity)
+        finish()
+    }
+
+    private fun askToLogout() {
+        val dialogBundle = ConfirmationDialogFragment.Bundler(
+            ASK_TO_LOGOUT_DIALOG_ID,
+            getString(R.string.logout_dialog_title),
+            getString(R.string.logout_dialog_message),
+            getString(R.string.logout_dialog_yes),
+            getString(android.R.string.cancel)
+        ).bundle()
+
+        navigation.navigate(R.id.action_profileFragment_to_ask_logout_dialog, dialogBundle)
     }
 
     private fun setOrCancelTimer(onDuty: Boolean) {
