+++ app/src/main/java/org/wildaid/ofish/ui/profile/ProfileFragment.kt
@@ -1,46 +1,49 @@
 package org.wildaid.ofish.ui.profile
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
+import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
+import androidx.lifecycle.Observer
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.Glide
 import kotlinx.android.synthetic.main.fragment_user_profile.*
-import org.wildaid.ofish.EventObserver
 import org.wildaid.ofish.R
 import org.wildaid.ofish.databinding.FragmentUserProfileBinding
-import org.wildaid.ofish.ui.base.ConfirmationDialogFragment
 import org.wildaid.ofish.ui.base.DIALOG_CLICK_EVENT
 import org.wildaid.ofish.ui.base.DialogButton
 import org.wildaid.ofish.ui.base.DialogClickEvent
+import org.wildaid.ofish.ui.base.REQUEST_PICK_IMAGE
 import org.wildaid.ofish.ui.home.ASK_CHANGE_DUTY_DIALOG_ID
 import org.wildaid.ofish.ui.home.ASK_TO_LOGOUT_DIALOG_ID
 import org.wildaid.ofish.ui.home.HomeActivityViewModel
-import org.wildaid.ofish.util.getViewModelFactory
+import org.wildaid.ofish.util.*
 
 class ProfileFragment : Fragment(R.layout.fragment_user_profile) {
-    private val activityViewModel: HomeActivityViewModel by viewModels { getViewModelFactory() }
+    private val activityViewModel: HomeActivityViewModel by activityViewModels { getViewModelFactory() }
     private var dataBinding: FragmentUserProfileBinding? = null
     private val navigation: NavController by lazy { findNavController() }
+    private lateinit var pendingImageUri: Uri
+
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
         initUI(view)
-        setObservers()
         subscribeToDialogEvents()
         showOfficerPhoto()
-//        dataBinding!!.switchDutyStatus.setOnCheckedChangeListener { _, isChecked ->
-//            if (isChecked) {
-//                activityViewModel.onDutyChanged(isChecked)
-//            } else {
-//                showDutyReport()
-//            }
-//        }
+        setObservers()
+    }
+
+    private fun setObservers() {
+        activityViewModel.onDutyStatusLiveData.observe(viewLifecycleOwner, Observer { dutyStatus ->
+            image_user.isEnabled = dutyStatus
+        })
     }
 
     private fun showOfficerPhoto() {
@@ -53,55 +56,40 @@ class ProfileFragment : Fragment(R.layout.fragment_user_profile) {
         }
     }
 
-    private fun setObservers() {
-        activityViewModel.userEventLiveData.observe(viewLifecycleOwner, EventObserver {
-            when (it) {
-                HomeActivityViewModel.UserEvent.AskUserLogoutEvent -> askToLogout()
-                HomeActivityViewModel.UserEvent.UserLogoutEvent -> onUserLoggedOut()
-                HomeActivityViewModel.UserEvent.BecomeNotAtSea -> navigateToPatrolSummary()
-                HomeActivityViewModel.UserEvent.AskDutyConfirmationEvent -> askToChangeDuty()
-            }
-        })
-    }
-
-    private fun navigateToPatrolSummary() {
-        navigation.navigate(R.id.action_profileFragment_to_patrolSummaryFragment2)
-    }
-
-    private fun onUserLoggedOut() {
-        navigation.navigate(R.id.action_profileFragment_to_login_activity)
-        activity?.finish()
-    }
-
-    private fun askToChangeDuty() {
-        val dialogBundle = ConfirmationDialogFragment.Bundler(
-            ASK_CHANGE_DUTY_DIALOG_ID,
-            getString(R.string.you_are_not_at_sea),
-            getString(R.string.change_status_to_at_sea),
-            getString(R.string.yes),
-            getString(android.R.string.cancel)
-        ).bundle()
-
-        navigation.navigate(R.id.ask_change_duty_dialog, dialogBundle)
-    }
-
-    private fun askToLogout() {
-        val dialogBundle = ConfirmationDialogFragment.Bundler(
-            ASK_TO_LOGOUT_DIALOG_ID,
-            getString(R.string.logout_dialog_title),
-            getString(R.string.logout_dialog_message),
-            getString(R.string.logout_dialog_yes),
-            getString(android.R.string.cancel)
-        ).bundle()
-
-        navigation.navigate(R.id.action_profileFragment_to_ask_logout_dialog, dialogBundle)
-    }
 
     private fun initUI(view: View) {
         dataBinding = FragmentUserProfileBinding.bind(view).apply {
+            this.lifecycleOwner = lifecycleOwner
             this.activityViewModel = this@ProfileFragment.activityViewModel
         }
+
+        dataBinding!!.imageUser.setOnClickListener { pickUserImage() }
     }
+
+    private fun pickUserImage() {
+        val pickImageIntent = createGalleryIntent()
+        pendingImageUri = createImageUri()
+        val takePhotoIntent = createCameraIntent(pendingImageUri)
+
+        val intentList: MutableList<Intent> = mutableListOf()
+        combineIntents(intentList, pickImageIntent)
+        combineIntents(intentList, takePhotoIntent)
+
+        val chooserIntent: Intent?
+        if (intentList.size > 0) {
+            chooserIntent = Intent.createChooser(
+                intentList.removeAt(intentList.size - 1),
+                getString(R.string.chose_image_source)
+            )
+            chooserIntent.putExtra(
+                Intent.EXTRA_INITIAL_INTENTS,
+                intentList.toTypedArray()
+            )
+
+            startActivityForResult(chooserIntent, REQUEST_PICK_IMAGE)
+        }
+    }
+
 
     private fun subscribeToDialogEvents() {
         val navStack = navigation.currentBackStackEntry!!
@@ -127,6 +115,8 @@ class ProfileFragment : Fragment(R.layout.fragment_user_profile) {
             }
             ASK_CHANGE_DUTY_DIALOG_ID -> {
                 if (event.dialogBtn == DialogButton.POSITIVE) activityViewModel.onDutyChanged(true)
+                if (event.dialogBtn == DialogButton.NEGATIVE) switch_duty_status.isChecked =
+                    activityViewModel.onDutyStatusLiveData.value!!
                 true
             }
             else -> false
